♻️ refactor: 예제 페이지들을 example 폴더로 이동
