+++ src/app/example/_components/PostList.tsx
@@ -1,20 +1,23 @@
 "use client";
 
 import { useQuery } from "@tanstack/react-query";
-import React from "react";
 
-import { getPosts } from "@/server/getPosts";
+import { getPosts } from "../_action/getPosts";
 
-const Posts = () => {
-  const { data, error, isFetched } = useQuery({
+const PostList = () => {
+  const { data, isLoading } = useQuery({
     queryKey: ["Posts"],
     queryFn: getPosts,
     staleTime: 10 * 1000,
   });
 
+  if (isLoading) {
+    return <h1>Loading...</h1>;
+  }
+
   return (
-    <ul className="grid grid-cols-1 gap-2">
-      {data.map((post: { id: string; title: string }) => (
+    <ul className="grid  w-full grid-cols-1 gap-2">
+      {data.slice(0, 10).map((post: { id: string; title: string }) => (
         <li
           key={post.id}
           className="rounded-md bg-gray-200 p-4 shadow-md dark:bg-gray-800"
@@ -26,4 +29,4 @@ const Posts = () => {
   );
 };
 
-export default Posts;
+export default PostList;
