+++ src/app/example/_components/ReactQueryServerActionExample.tsx
@@ -1,0 +1,24 @@
+import {
+  dehydrate,
+  HydrationBoundary,
+  QueryClient,
+} from "@tanstack/react-query";
+
+import { getPosts } from "../_action/getPosts";
+import PostList from "./PostList";
+
+const ReactQueryServerActionExample = async () => {
+  const queryClient = new QueryClient();
+  await queryClient.prefetchQuery({
+    queryKey: ["Posts"],
+    queryFn: getPosts,
+  });
+
+  return (
+    <HydrationBoundary state={dehydrate(queryClient)}>
+      <PostList />
+    </HydrationBoundary>
+  );
+};
+
+export default ReactQueryServerActionExample;
