--- src/app/zustand/page.tsx
@@ -1,11 +1,0 @@
-import ZustnadExample from "@/components/ZustnadExample";
-
-const page = () => {
-  return (
-    <main>
-      <ZustnadExample />
-    </main>
-  );
-};
-
-export default page;
